- 형변환 오류 뜨던거 다 고침 - RenderObject Create 사용하지 않고 AObject에서 unique_ptr 사용해서 코드 짜봄

+++ Program/Engine/Private/Model.cpp
@@ -389,7 +389,7 @@ HRESULT CModel::Ready_Materials(const char * pModelFilePath)
 			strcat_s(szFullPath, szExt);
 
 			_tchar		szTextureFullPath[MAX_PATH] = TEXT("");
-			MultiByteToWideChar(CP_ACP, 0, szFullPath, strlen(szFullPath), szTextureFullPath, MAX_PATH);
+			MultiByteToWideChar(CP_ACP, 0, szFullPath, (int)strlen(szFullPath), szTextureFullPath, (int)MAX_PATH);
 
 			pMeshMaterialDesc->pTexture[j] = CTexture::Create(m_pDevice, m_pDeviceContext, szTextureFullPath);
 			if (nullptr == pMeshMaterialDesc->pTexture[j])
